requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9.2
dateparser==0.7.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.41.0
django-bootstrap-form==3.4
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.9.0
django-filter==2.2.0
django-googlesearch==0.2
django-js-asset==1.2.2
django-social-share==1.4.0
django-taggit==1.2.0
docutils==0.16
Flask==1.1.2
get==2019.4.13
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
Markdown==3.2.1
MarkupSafe==1.1.1
mock==4.0.2
mysqlclient
note==0.5.2
oauthlib==3.1.0
Pillow==7.1.1
post==2019.4.13
psycopg2==2.8.5
public==2019.4.13
pycparser==2.20
PyJWT==1.7.1
pymongo==3.10.1
python-daemon==2.2.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pyzmq==19.0.0
query-string==2019.4.13
regex==2020.4.4
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
tornado==6.0.4
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.1
